% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRIER2-package.R
\docType{package}
\name{BRIER2-package}
\alias{BRIER2}
\alias{BRIER2-package}
\title{BRIER2: Bregman-Divergence-Based Regularized Integrative Estimator in Regression}
\description{
BRIER2 is a transfer learning framework for improving genetic risk prediction in a target dataset by leveraging information from external, well-powered cohorts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://zrayw.github.io/BRIER2/}
}

}
\author{
\strong{Maintainer}: Ruiwen Zhou \email{zrayw@umich.edu}

}
\keyword{internal}
