% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{g_link}
\alias{g_link}
\title{Forward link function}
\usage{
g_link(mu, family)
}
\arguments{
\item{mu}{A numeric vector or matrix of mean values.}

\item{family}{A string: "gaussian", "binomial", or "poisson".}
}
\value{
A numeric vector or matrix of linear predictors on the link scale.
}
\description{
Maps the mean (mu) to the linear predictor (eta) under a specified
GLM family.
}
\seealso{
\code{\link{ginv_link}}
}
