% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\name{getLDB}
\alias{getLDB}
\title{Path to a bundled Berisa-Pickrell LD block file}
\usage{
getLDB(ancestry = c("AFR", "EAS", "EUR"), build = c("hg38", "hg19"))
}
\arguments{
\item{ancestry}{Character. One of \code{"AFR"}, \code{"EAS"}, or
\code{"EUR"}.}

\item{build}{Character. One of \code{"hg19"} (GRCh37) or \code{"hg38"}
(GRCh38). Default \code{"hg38"}.}
}
\value{
Absolute path to the requested BED file.
}
\description{
Returns the file system path to a packaged BED file that describes
approximately-independent LD blocks from Berisa and Pickrell (2016).
The package ships LD block coordinates for three continental ancestries
(AFR, EAS, EUR) in two genome builds (hg19, hg38).
}
\details{
These files can be used directly as the \code{blocks} argument to
\code{\link{LD_blocks}} or supplied to downstream LD-aware functions
such as \code{\link{calLD}}.
}
\examples{
\dontrun{
bed_path <- getLDB("EUR", "hg38")
blocks <- read.table(bed_path, header = T)
head(blocks)
}

}
\references{
Berisa, T. and Pickrell, J. K. (2016). Approximately independent
linkage disequilibrium blocks in human populations.
\emph{Bioinformatics}, 32(2), 283-285.

Original hg19 BED files:
\url{https://bitbucket.org/nygcresearch/ldetect-data}
}
\seealso{
\code{\link{LD_blocks}}, \code{\link{calLD}}
}
