% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LD_sigma}
\alias{LD_sigma}
\title{Block-Structured LD Calculation}
\usage{
LD_sigma(X, blk, tau)
}
\arguments{
\item{X}{N x p standardized genotype matrix.}

\item{blk}{Vector of block boundaries of length K+1 (0-based).}

\item{tau}{Soft-threshold parameter.}
}
\value{
p x p sparse LD matrix (arma::sp_mat).
}
\description{
Calculate LD matrix within blocks using soft-thresholding.
}
