% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{standardize_X}
\alias{standardize_X}
\title{Standardize a design matrix}
\usage{
standardize_X(Xr)
}
\arguments{
\item{Xr}{An n x p numeric matrix.}
}
\value{
A list with the following elements:
\describe{
\item{standardized}{The standardized n x p matrix.}
\item{center}{A numeric vector of length p of column means.}
\item{scale}{A numeric vector of length p of column standard deviations.}
}
}
\description{
Center and scale each column of X to have mean 0 and standard deviation 1.
Constant columns (sd = 0) are centered but not scaled.
}
